+++ src/Framework/Dispatcher.php
@@ -43,49 +43,51 @@ class Dispatcher
         $this->hook->trigger(self::HOOK_DISPATCH_BEFORE, $route);
         $controller = null;
 
-        /* Start output buffering and set $content to null */
+        // Start output buffering and set $content to null
         $content = null;
         $this->response->startOutputBuffer();
 
-        /* Build the parameters array */
-        $parameters = [$route];
+        // Build the parameters array
+        $parameters = [];
         foreach ($route->getValues() as $value) {
             $parameters[] = $value;
         }
 
-        /* Call the relevant code */
-        if ($route->controller && $route->action) {
-            $controller = \Parable\DI\Container::get($route->controller);
-            $content = $controller->{$route->action}(...$parameters);
-        } elseif ($route->callable) {
-            $call = $route->callable;
-            $content = $call(...$parameters);
+        // Call the relevant code
+        if ($route->hasControllerAndAction()) {
+            $controller = \Parable\DI\Container::get($route->getController());
+            $content = $controller->{$route->getAction()}(...$parameters);
+        } elseif ($route->hasCallable()) {
+            $callable = $route->getCallable();
+            $content = $callable(...$parameters);
         }
 
-        /* Try to get the relevant view */
-        $templateFile = null;
-        if ($route->template) {
-            $templateFile = $this->path->getDir($route->template);
-        } else {
-            if ($controller) {
-                $reflection = new \ReflectionClass($controller);
-                $controllerName = str_replace('\\', '/', $reflection->getName());
-                $controllerName = str_replace('Controller/', '', $controllerName);
-                $templateFile = $this->path->getDir(
-                    "app/View/{$controllerName}/{$route->action}.phtml"
-                );
+        // If the route has no template path, attempt to build one based on controller/action.phtml
+        if (!$route->hasTemplatePath() && $route->hasControllerAndAction()) {
+            $reflection = new \ReflectionClass($controller);
+            $controllerName = str_replace('\\', '/', $reflection->getName());
+            $controllerName = str_replace('Controller/', '', $controllerName);
+
+            $templatePathGenerated = $this->path->getDir(
+                "app/View/{$controllerName}/{$route->getAction()}.phtml"
+            );
+
+            if (file_exists($templatePathGenerated)) {
+                $route->setTemplatePath($templatePathGenerated);
             }
         }
 
-        if ($templateFile && file_exists($templateFile)) {
-            $this->view->setTemplatePath($templateFile);
+        // And check again, now that we might have a magic template path
+        if ($route->hasTemplatePath()) {
+            $templatePath = $this->path->getDir($route->getTemplatePath());
+            $this->view->setTemplatePath($templatePath);
             $this->view->render();
         }
 
-        /* Get the output buffer content and check if $content holds anything. If so, append it to the $bufferContent */
+        // Get the output buffer content and check if $content holds anything. If so, append it to the $bufferContent
         $content = $this->response->returnOutputBuffer() . $content;
 
-        /* And append the content to the response object */
+        // And append the content to the response object
         $this->response->appendContent($content);
 
         $this->hook->trigger(self::HOOK_DISPATCH_AFTER, $route);